docs/requirements.txt
# Sphinx documentation requirements
sphinx>=7.0.0
karma-sphinx-theme>=0.0.8
myst-parser>=2.0.0
sphinx-copybutton>=0.5.0
sphinx-tabs>=3.4.0
sphinx-design>=0.5.0
sphinx-autobuild>=2021.3.14
linkify-it-py>=2.0.0
PyYAML>=6.0.0

# Additional documentation tools
sphinx-rtd-theme>=1.3.0
furo>=2023.9.10
sphinx-book-theme>=1.0.0

# Markdown support
recommonmark>=0.7.1
sphinx-markdown-tables>=0.0.17

# Code highlighting
pygments>=2.16.0

# Link checking
sphinx-external-toc>=0.3.0
sphinx-sitemap>=2.5.0
